Refactor toggle text so it can be used outside of Storyblok (eg. in methodology page)
